+++ core/src/main/java/de/unistuttgart/iaas/amyassist/amy/naturallang/GrammarParser.java
@@ -21,7 +21,7 @@
  * For more information see notice.md
  */
 
-package de.unistuttgart.iaas.amyassist.amy.core;
+package de.unistuttgart.iaas.amyassist.amy.naturallang;
 
 import java.util.ArrayList;
 import java.util.HashMap;
