+++ core/src/main/java/de/unistuttgart/iaas/amyassist/amy/naturallang/PluginGrammarInfo.java
@@ -21,7 +21,7 @@
  * For more information see notice.md
  */
 
-package de.unistuttgart.iaas.amyassist.amy.core;
+package de.unistuttgart.iaas.amyassist.amy.naturallang;
 
 import java.util.List;
 import java.util.Set;
