+++ core/src/test/java/de/unistuttgart/iaas/amyassist/amy/naturallang/GrammarParserTest.java
@@ -21,10 +21,13 @@
  * For more information see notice.md
  */
 
-package de.unistuttgart.iaas.amyassist.amy.core;
+package de.unistuttgart.iaas.amyassist.amy.naturallang;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import de.unistuttgart.iaas.amyassist.amy.naturallang.GrammarParser;
+
 import static org.hamcrest.MatcherAssert.*;
 import static org.hamcrest.Matchers.*;
